@@ -10,20 +10,13 @@
 		"format": "pnpm rome check . --apply-unsafe && pnpm rome format . --write"
 	},
 	"dependencies": {
-		"@vercel/analytics": "1.2.2",
-		"@vercel/edge-config": "0.4.1",
-		"contentlayer": "^0.3.4",
-		"date-fns": "^3.6.0",
+		"@vercel/analytics": "^1.2.2",
 		"gsap": "^3.12.5",
-		"next": "13.4.16",
-		"next-contentlayer": "^0.3.4",
+		"next": "14.1.4",
+		"next-mdx-remote": "^4.4.1",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
-		"react-wrap-balancer": "^1.1.0",
-		"rehype-autolink-headings": "^7.1.0",
-		"rehype-pretty-code": "^0.10.0",
-		"rehype-slug": "^6.0.0",
-		"remark-gfm": "^3.0.1",
+		"sugar-high": "^0.6.0",
 		"tailwindcss": "^3.4.1"
 	},
 	"devDependencies": {
