@@ -54,7 +54,7 @@ export default function RootLayout({ children }) {
 	return (
 		<html lang="en">
 			<body className={inter.className}>
-				<main className="flex flex-col min-h-screen bg-gradient-to-tl from-black via-zinc-600/20 to-black">
+				<main className="flex flex-col min-h-screen bg-gradient-to-tl from-black via-zinc-600/20 to-black overflow-x-hidden">
 					<Particles
 						className="absolute inset-0 -z-10 animate-fade-in"
 						quantity={100}
