@@ -1,9 +1,9 @@
-import { allBlogs } from "contentlayer/generated";
+import { getBlogPosts } from "@/db/blog";
 
 export default async function sitemap() {
-	const blogs = allBlogs.map((post) => ({
+	const blogs = getBlogPosts().map((post) => ({
 		url: `https://justinjdaniel.com/blog/${post.slug}`,
-		lastModified: post.publishedAt,
+		lastModified: post.metadata.publishedAt,
 	}));
 
 	const routes = ["", "/blog", "/projects", "/uses"].map((route) => ({
